requirements.txt
alembic==1.0.8
aniso8601==6.0.0
atomicwrites==1.3.0
attrs==19.1.0
bcrypt==3.1.6
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
coverage==4.5.3
coveralls==1.7.0
docopt==0.6.2
Flask==1.0.2
Flask-API==1.1
Flask-Bcrypt==0.7.1
Flask-Migrate==2.4.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.8
MarkupSafe==1.1.1
more-itertools==7.0.0
pluggy==0.9.0
psycopg2==2.8.1
py==1.8.0
pycparser==2.19
PyJWT==1.7.1
pytest==4.4.1
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
pytz==2019.1
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.2
urllib3==1.25.3
Werkzeug==0.15.2
